Separated createObserver.js file
