+++ lib/index.js
@@ -12,10 +12,12 @@ var inherits = require('util').inherits;
 var EventEmitter = require('events').EventEmitter;
 var Decoder = require('string_decoder').StringDecoder;
 
+var createObserver = require('./createObserver');
+var isFunction = require('./isFunction');
 var IntMap = require('./intMap');
+var ObservableSubscription = require('./observableSubscription');
 var Queue = require('./queue');
 var ReadableProxy = require('./readableProxy');
-var ObservableSubscription = require('./observableSubscription');
 
 // Create quick slice reference variable for speed
 var slice = Array.prototype.slice;
@@ -312,9 +314,7 @@ _.isUndefined = function (x) {
     return typeof x === 'undefined';
 };
 
-_.isFunction = function (x) {
-    return typeof x === 'function';
-};
+_.isFunction = isFunction;
 
 _.isObject = function (x) {
     return typeof x === 'object' && x !== null;
@@ -2100,57 +2100,6 @@ addMethod('toPromise', function (PromiseCtor) {
     });
 });
 
-/**
- * Coerce observer callbacks into observer object or return observer object
- * if already created. Will throw an error if both an object and callback args
- * are provided.
- *
- * @id createObserver
- * @param {function|object} onNext Function to receive new values or observer
- * @param {function} [onError] Optional callback to receive errors.
- * @param {function} [onComplete] Optional callback when stream completes
- * @return {object} Observer object with next, error, and complete methods
- * @private
- *
- * createObserver(
- *     function (x) { console.log(x); },
- *     function (err) { console.error(err); },
- *     function () { console.log('done'); }
- * )
- *
- * createObserver(
- *     null,
- *     null,
- *     function () { console.log('done'); }
- * )
- *
- * createObserver({
- *     next: function (x) { console.log(x); },
- *     error: function (err) { console.error(err); },
- *     complete: function () { console.log('done'); }
- * })
- */
-function createObserver (onNext, onError, onComplete) {
-    var isObserver = onNext && !_.isFunction(onNext) && typeof onNext === 'object';
-
-    // ensure if we have an observer that we don't also have callbacks. Users
-    // must choose one.
-    if (isObserver && (onError || onComplete)) {
-        throw new Error('Subscribe requires either an observer object or optional callbacks.');
-    }
-
-    // onNext is actually an observer
-    if (isObserver) {
-        return onNext;
-    }
-
-    // Otherwise create an observer object
-    return {
-        next: onNext,
-        error: onError,
-        complete: onComplete,
-    };
-}
 
 /**
  * Consumes values using the Observable subscribe signature. Unlike other
